restructured the python services portion of the project; included a readme for connecting to the venv environment
