+++ .gitignore
@@ -413,6 +413,11 @@ dist
 __pycache__/
 *.py[cod]
 *$py.class
+*.so
+.Python
+venv/
+ENV/
+.env
 
 # C extensions
 *.so
