+++ python_services/requirements.txt
@@ -1,0 +1,6 @@
+psycopg2-binary==2.9.9
+python-dotenv==1.0.1
+pandas==2.2.0
+numpy==1.26.3
+matplotlib==3.8.2
+scikit-learn==1.4.0
